refactor(extension): simplify initialization tests
